offboarding-agent/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
supabase>=2.0.0
click>=8.1.7
